Add DELTA_BYTE_ARRAY decoding
This adds decoder for DELTA_BYTE_ARRAY decoding. To implement this, I
found that I also need to fix the `MemoryPool` class, as all the public
APIs of it cannot take `&mut self`. Otherwise, calls such as

  self.memory_pool.consume(data)

will return result with same lifetime as `self` (which is definitely
shorter than `self.memory_pool`s). The reason for this is because Rust
is invariant on lifetime of mutable references, but contravariant on
lifetime of immutable references.

Test is needed for the DELTA_BYTE_ARRAY decoder.