Refactor Type & add SchemaDescriptor
At the moment types, including both primitive type & group type, are
implemented using trait. However, this is not so nice as:
1. given a trait it is hard to know which concrete type it is, and is
hard to downcast the trait to that specific type.
2. trait has to be wrapped with Box, which incurs extra costs.

This changes the implementation to enum, with which we can do pattern
matching on the type and "look into" the specific type. Also, this
replaces the visitor pattern with pattern matching.

Also, this adds a SchemaDescriptor asides from the existing
ColumnDescriptor. It also adds mechanism to construct ColumnDescriptors
from a top-level schema. Test is needed for this process.