Remove lifetime param for Column & SchemaDescriptor
Lifetime for these two is hard to deal with because currently there's no
good way for a field to self-reference other fields inside the same
struct. Because of this, the struct A who stores the descriptors also
need to have a lifetime parameter, and so does the struct who stores
struct A, and so on. This causes a cascading effect and there's no way
to "localize" the lifetime.

Here we replace the `&'a Type` in the descriptors with `Rc<Type>`. This
is reasonable since we needs multiple aliases given a `Type` instance,
and that we seldomly need write access to the instance.