Use Rc<Vec<u8>> for decoders (instead of slices)
The current approach of using slice for byte arrays doesn't seem to
work, since it requires a continus stack frames for the lifetime
parameters, and therefore the eventual decoded values can not outlive
the "root" stack frame where the bytes originated from.

Another issue is with column value reader, as we need to first read the
page and then read values in batches. However, we cannot have decoder
and page reader both as fields for the value reader, since the former
contains reference to the latter. This is again an issue of
self-reference, which Rust doesn't handle right now without 3rd party
crates.