Add BytePtr which is a slice on Rc<Vec<u8>>
In previous commit, many places we need to get a slice on the
Rc<Vec<u8>> type. As result, we added `offset` and `len` to many
function parameters, which is not so nice.

This changes `BytePtr`, which currently is a alias to `Rc<Vec<u8>>`, to
a struct that is a tuple of {Rc<Vec<u8>>, start, len}, which is equipped
with several operations such as getting a slice (another `BytePtr`) from
`BytePtr`.