Implement Seek for BitWriter and add advance_num_bytes
The `get_next_byte_ptr` is a little inconvenient to use since it
requires the whole permission for self to be carved out while the result
slice is in scope. This commit added a new method `advance_num_bytes`
that is similar to `get_next_byte_ptr`, but simply return the old offset
before advancing. Also implemented `std::io::Seek` for `BitWriter` so
that the caller can then call `Seek` to come back and fill the
designated bytes.