Add ctor to RawRleEncoder & BitWriter to accept allocated buffer
This adds extra ctor `new_from_buf` for `RawRleEncoder` and `BitWriter`
which accept an allocated buffer (`Vec<u8>`) as input. This is necessary
since sometimes we need to pass a buffer around.

Also added `clear` for the above two which clears the internal state so
the buffer can be reused.

Lastly, changed `consume` to return `Vec<u8>` instead of
`ByteBufferPtr`. The former is more flexible.