Add LevelEncoder and LevelDecoder
The current decoding::RleDecoder is only used for decoding rep/def
levels. However, it can only decode i32 while levels are encoded as i16.
Also, it doesn't fit well within the Decoder trait.

This commit creates two separate structs: LevelEncoder and LevelDecoder,
to replace the above. This is similar to the C++ implementation. These
two are thin wrappers on the existing RawRleEncoder and RawRleDecoder.

Also did a little bit refactoring:
1. BitWriter::consume() and RawRleEncoder::consume() no longer consume
   "this". Instead they just replace the underlying buffer.
2. move RawRleEncoder::min_buffer_size and
   RawRleEncoder::max_buffer_size to the front of struct definition
   since they are "static" functions.