BitReader: change return type from Result to Option
Currently most of the BitReader functions return Result type in the case
there's no more data available in the stream. This is not so appropriate
as this case should not be considered as error. This changes Result to
Option.

There are some cases though, that we should return Err if there's no
more data available (e.g., because invariant is broken). For instance,
in Decoder. Currently we return ParquetErr::General for these cases.
This commit defines a new type of error: ParquetErr::EOF, to replace
them.