Fix buffer too small problem for DictEncoder
In the DictEncoder::write_indices(), the output buffer is allocated with
buffer size RleEncoder::min_buffer_size, which is the max buffer size
for a single RLE/bit-packed run. This is too small if the values to be
flushed need multiple runs. The issue can be revealed by setting
TEST_SET_SIZE to a larger value such as 1024.