Replace x86intrin with stable SIMD (#118)
Currently the SIMD feature requires the x86intrin crate and a compile
time flag `RUSTFLAGS="-C target-feature=+sse4.2"`, which is not
convenient. This replaces the x86intrin dependency with recently
stablized SIMD features, and also replaces the `cfg(target_feature)`
flag with `is_x86_feature_detected`, which dynamically determines
whether SSE 4.2 is supported or not.

Fixes #113.