Generic reader (#142)
Generalized FileReader in order to allow multiple inner readers. This
will allow to be able to parse from an in-memory buffer, for example
(via Cursor<&[u8]>, for example).

Note that there are still some issues because of 'static lifetimes (I
guess that it's still due to some type referencing `DataType`), which
limits a little bit how it may be used. For example, this example does
not work at the moment:

```rust
    let v = vec![0, 1, 2, 3];
    let c = Cursor::new(v.as_ref());
    let serialized = SerializedFileReader::new(c).unwrap();

    // compiler refuses to compile due to `v does not life enough`
    let iterator = serialized.get_row_iter(None).unwrap();
```